refactor(halley-wl): split root, render, and workspace lifecycle subsystems
- move exit confirmation, screenshot flow, and camera-oriented behavior out of central compositor hubs into dedicated modules/controllers
- extract screenshot pointer handling into its own input path and update keyboard/pointer plumbing to match
- split render/window into focused submodules and separate frame runtime orchestration from frame drawing
- replace monolithic workspace/lifecycle with a module tree and adjust focus, cluster, carry, spawn, and action wiring around it
- simplify small support areas like spatial helpers and module boundaries without intended behavior changes
