+++ crates/halley-wl/src/input/pointer/mod.rs
@@ -3,6 +3,7 @@ pub(crate) mod button;
 pub(crate) mod focus;
 pub(crate) mod motion;
 pub(crate) mod resize;
+mod screenshot;
 
 pub(crate) use resize::{
     active_node_screen_rect, active_node_surface_transform_screen_details,
