+++ crates/halley-wl/src/render/window/mod.rs
@@ -33,6 +33,15 @@ use super::clipped_surface::ClippedSurfaceRenderElement;
 use super::offscreen::render_surface_tree_to_texture;
 use super::utils::{sync_node_size_from_surface, world_to_screen};
 
+mod offscreen;
+mod stack;
+
+use offscreen::world_to_screen_for_view;
+pub(crate) use offscreen::{
+    capture_closing_window_animation, prewarm_visible_active_window_offscreen_caches,
+};
+use stack::{build_stack_transition_plan, clone_stack_window_unit_for_pose, stack_draw_order_map};
+
 type SurfaceElement =
     smithay::backend::renderer::element::surface::WaylandSurfaceRenderElement<GlesRenderer>;
 render_elements! {
@@ -142,316 +151,6 @@ impl StackWindowDrawUnit {
             offscreen_textures: Vec::new(),
         }
     }
-}
-
-#[derive(Clone, Copy)]
-struct StackTransitionPose {
-    center: Vec2,
-    size: Vec2,
-    alpha: f32,
-    draw_order: i32,
-}
-
-struct StackTransitionExtraInstance {
-    node_id: NodeId,
-    pose: StackTransitionPose,
-}
-
-struct StackTransitionPlan {
-    poses: HashMap<NodeId, StackTransitionPose>,
-    extra_instances: Vec<StackTransitionExtraInstance>,
-}
-
-fn lerp_f32(a: f32, b: f32, t: f32) -> f32 {
-    a + (b - a) * t
-}
-
-fn lerp_vec2(a: Vec2, b: Vec2, t: f32) -> Vec2 {
-    Vec2 {
-        x: lerp_f32(a.x, b.x, t),
-        y: lerp_f32(a.y, b.y, t),
-    }
-}
-
-fn rect_center(rect: Rect) -> Vec2 {
-    Vec2 {
-        x: rect.x + rect.w * 0.5,
-        y: rect.y + rect.h * 0.5,
-    }
-}
-
-fn rect_size(rect: Rect) -> Vec2 {
-    Vec2 {
-        x: rect.w.max(1.0),
-        y: rect.h.max(1.0),
-    }
-}
-
-fn stack_draw_order_map(front_to_back: &[NodeId]) -> HashMap<NodeId, i32> {
-    let len = front_to_back.len() as i32;
-    front_to_back
-        .iter()
-        .enumerate()
-        .map(|(index, &node_id)| (node_id, len - index as i32 - 1))
-        .collect()
-}
-
-fn build_stack_transition_plan(
-    st: &Halley,
-    monitor: &str,
-    transition: &crate::render::state::StackCycleTransitionSnapshot,
-) -> Option<StackTransitionPlan> {
-    let custom_source_rects = transition.source_rects.is_some();
-    let old_rects = transition
-        .source_rects
-        .clone()
-        .or_else(|| st.stack_layout_rects_for_members(monitor, &transition.old_visible))?;
-    let new_rects = st.stack_layout_rects_for_members(monitor, &transition.new_visible)?;
-    let t = ease_in_out_cubic(transition.progress);
-    let draw_orders = stack_draw_order_map(&transition.new_visible);
-    let topmost_order = transition.new_visible.len() as i32 + 1;
-    let old_top = transition.old_visible.first().copied();
-    let old_bottom = transition.old_visible.last().copied();
-    let new_top = transition.new_visible.first().copied();
-    let wrapped_same_set = !custom_source_rects
-        && transition.old_visible.len() == transition.new_visible.len()
-        && transition
-            .old_visible
-            .iter()
-            .all(|id| transition.new_visible.contains(id));
-    let wrapped_node = match transition.direction {
-        ClusterCycleDirection::Next
-            if wrapped_same_set && old_top == transition.new_visible.last().copied() =>
-        {
-            old_top
-        }
-        ClusterCycleDirection::Prev if wrapped_same_set && old_bottom == new_top => old_bottom,
-        _ => None,
-    };
-
-    let mut ids = transition.old_visible.clone();
-    for &node_id in &transition.new_visible {
-        if !ids.contains(&node_id) {
-            ids.push(node_id);
-        }
-    }
-
-    let mut poses = HashMap::new();
-    let mut extra_instances = Vec::new();
-    for node_id in ids {
-        if wrapped_node == Some(node_id)
-            && let (Some(old_rect), Some(new_rect)) = (
-                old_rects.get(&node_id).copied(),
-                new_rects.get(&node_id).copied(),
-            )
-        {
-            let canonical_pose = match transition.direction {
-                ClusterCycleDirection::Next => StackTransitionPose {
-                    center: rect_center(new_rect),
-                    size: rect_size(new_rect),
-                    alpha: t,
-                    draw_order: draw_orders.get(&node_id).copied().unwrap_or_default(),
-                },
-                ClusterCycleDirection::Prev => {
-                    let end_center = rect_center(new_rect);
-                    let mut start_center = end_center;
-                    start_center.x -= new_rect.w * 0.55;
-                    StackTransitionPose {
-                        center: lerp_vec2(start_center, end_center, t),
-                        size: rect_size(new_rect),
-                        alpha: t,
-                        draw_order: topmost_order,
-                    }
-                }
-            };
-            poses.insert(node_id, canonical_pose);
-
-            if matches!(transition.direction, ClusterCycleDirection::Next) {
-                let mut end_center = rect_center(old_rect);
-                end_center.x -= old_rect.w * 0.55;
-                extra_instances.push(StackTransitionExtraInstance {
-                    node_id,
-                    pose: StackTransitionPose {
-                        center: lerp_vec2(rect_center(old_rect), end_center, t),
-                        size: rect_size(old_rect),
-                        alpha: 1.0 - t,
-                        draw_order: topmost_order,
-                    },
-                });
-            }
-            continue;
-        }
-
-        let pose = match (
-            old_rects.get(&node_id).copied(),
-            new_rects.get(&node_id).copied(),
-        ) {
-            (Some(old_rect), Some(new_rect)) => StackTransitionPose {
-                center: lerp_vec2(rect_center(old_rect), rect_center(new_rect), t),
-                size: lerp_vec2(rect_size(old_rect), rect_size(new_rect), t),
-                alpha: 1.0,
-                draw_order: draw_orders.get(&node_id).copied().unwrap_or_default(),
-            },
-            (Some(old_rect), None) => {
-                let mut end_center = rect_center(old_rect);
-                let draw_order = match transition.direction {
-                    ClusterCycleDirection::Next if Some(node_id) == old_top => {
-                        end_center.x -= old_rect.w * 0.55;
-                        topmost_order
-                    }
-                    ClusterCycleDirection::Prev if Some(node_id) == old_bottom => {
-                        end_center.x += old_rect.w * 0.08;
-                        end_center.y += old_rect.h * 0.04;
-                        0
-                    }
-                    _ => 0,
-                };
-                StackTransitionPose {
-                    center: lerp_vec2(rect_center(old_rect), end_center, t),
-                    size: rect_size(old_rect),
-                    alpha: 1.0 - t,
-                    draw_order,
-                }
-            }
-            (None, Some(new_rect)) => {
-                let end_center = rect_center(new_rect);
-                let mut start_center = end_center;
-                let draw_order = match transition.direction {
-                    ClusterCycleDirection::Prev if Some(node_id) == new_top => {
-                        start_center.x -= new_rect.w * 0.55;
-                        topmost_order
-                    }
-                    _ => draw_orders.get(&node_id).copied().unwrap_or_default(),
-                };
-                StackTransitionPose {
-                    center: lerp_vec2(start_center, end_center, t),
-                    size: rect_size(new_rect),
-                    alpha: t,
-                    draw_order,
-                }
-            }
-            (None, None) => continue,
-        };
-        poses.insert(node_id, pose);
-    }
-
-    Some(StackTransitionPlan {
-        poses,
-        extra_instances,
-    })
-}
-
-fn transform_rect_about_center(
-    x: i32,
-    y: i32,
-    w: i32,
-    h: i32,
-    from_center: (f32, f32),
-    to_center: (f32, f32),
-    scale_x: f32,
-    scale_y: f32,
-) -> (i32, i32, i32, i32) {
-    let rect_center_x = x as f32 + w as f32 * 0.5;
-    let rect_center_y = y as f32 + h as f32 * 0.5;
-    let new_center_x = to_center.0 + (rect_center_x - from_center.0) * scale_x;
-    let new_center_y = to_center.1 + (rect_center_y - from_center.1) * scale_y;
-    let new_w = (w as f32 * scale_x).round().max(1.0) as i32;
-    let new_h = (h as f32 * scale_y).round().max(1.0) as i32;
-    (
-        (new_center_x - new_w as f32 * 0.5).round() as i32,
-        (new_center_y - new_h as f32 * 0.5).round() as i32,
-        new_w,
-        new_h,
-    )
-}
-
-fn clone_stack_window_unit_for_pose(
-    st: &Halley,
-    size: Size<i32, Physical>,
-    unit: &StackWindowDrawUnit,
-    from_pose: StackTransitionPose,
-    to_pose: StackTransitionPose,
-) -> Option<StackWindowDrawUnit> {
-    let (from_cx, from_cy) =
-        world_to_screen(st, size.w, size.h, from_pose.center.x, from_pose.center.y);
-    let (to_cx, to_cy) = world_to_screen(st, size.w, size.h, to_pose.center.x, to_pose.center.y);
-    let scale_x = (to_pose.size.x / from_pose.size.x.max(1.0)).max(0.01);
-    let scale_y = (to_pose.size.y / from_pose.size.y.max(1.0)).max(0.01);
-
-    let border_rect = unit.border_rect.as_ref().cloned().map(|mut rect| {
-        let (x, y, w, h) = transform_rect_about_center(
-            rect.x,
-            rect.y,
-            rect.w,
-            rect.h,
-            (from_cx as f32, from_cy as f32),
-            (to_cx as f32, to_cy as f32),
-            scale_x,
-            scale_y,
-        );
-        rect.x = x;
-        rect.y = y;
-        rect.w = w;
-        rect.h = h;
-        rect.inner_w = w as f32;
-        rect.inner_h = h as f32;
-        rect.alpha *= to_pose.alpha.clamp(0.0, 1.0);
-        rect
-    });
-
-    let offscreen_textures = unit
-        .offscreen_textures
-        .iter()
-        .cloned()
-        .map(|mut tex| {
-            let (dst_x, dst_y, dst_w, dst_h) = transform_rect_about_center(
-                tex.dst_x,
-                tex.dst_y,
-                tex.dst_w,
-                tex.dst_h,
-                (from_cx as f32, from_cy as f32),
-                (to_cx as f32, to_cy as f32),
-                scale_x,
-                scale_y,
-            );
-            let (clip_x, clip_y, clip_w, clip_h) = transform_rect_about_center(
-                tex.clip_x,
-                tex.clip_y,
-                tex.clip_w,
-                tex.clip_h,
-                (from_cx as f32, from_cy as f32),
-                (to_cx as f32, to_cy as f32),
-                scale_x,
-                scale_y,
-            );
-            tex.dst_x = dst_x;
-            tex.dst_y = dst_y;
-            tex.dst_w = dst_w;
-            tex.dst_h = dst_h;
-            tex.clip_x = clip_x;
-            tex.clip_y = clip_y;
-            tex.clip_w = clip_w;
-            tex.clip_h = clip_h;
-            tex.geo_offset_x *= scale_x;
-            tex.geo_offset_y *= scale_y;
-            tex.geo_w *= scale_x;
-            tex.geo_h *= scale_y;
-            tex.alpha *= to_pose.alpha.clamp(0.0, 1.0);
-            tex
-        })
-        .collect::<Vec<_>>();
-
-    if border_rect.is_none() && offscreen_textures.is_empty() {
-        return None;
-    }
-
-    Some(StackWindowDrawUnit {
-        node_id: unit.node_id,
-        draw_order: to_pose.draw_order,
-        border_rect,
-        active_elements: Vec::new(),
-        offscreen_textures,
-    })
 }
 
 fn rect_from_local_geometry(
@@ -595,306 +294,6 @@ fn offscreen_visual_crop_and_dst(
         clip.size.w,
         clip.size.h,
     )
-}
-
-fn render_view_for_monitor(st: &Halley, monitor: &str) -> (Vec2, Vec2, Vec2) {
-    if st.model.monitor_state.current_monitor == monitor {
-        return (
-            st.model.viewport.center,
-            st.model.viewport.size,
-            st.model.zoom_ref_size,
-        );
-    }
-
-    st.model
-        .monitor_state
-        .monitors
-        .get(monitor)
-        .map(|space| {
-            (
-                space.viewport.center,
-                space.viewport.size,
-                space.zoom_ref_size,
-            )
-        })
-        .unwrap_or((
-            st.model.viewport.center,
-            st.model.viewport.size,
-            st.model.zoom_ref_size,
-        ))
-}
-
-fn world_to_screen_for_view(
-    view_center: Vec2,
-    view_size: Vec2,
-    output_w: i32,
-    output_h: i32,
-    x: f32,
-    y: f32,
-) -> (i32, i32) {
-    let vw = view_size.x.max(1.0);
-    let vh = view_size.y.max(1.0);
-    let nx = ((x - view_center.x) / vw) + 0.5;
-    let ny = ((y - view_center.y) / vh) + 0.5;
-
-    (
-        (nx * output_w as f32).round() as i32,
-        (ny * output_h as f32).round() as i32,
-    )
-}
-
-pub(crate) fn capture_closing_window_animation(
-    st: &Halley,
-    monitor: &str,
-    node_id: NodeId,
-) -> Option<(Option<ActiveBorderRect>, Vec<OffscreenNodeTexture>)> {
-    let node = st.model.field.node(node_id)?;
-    let cache = st.ui.render_state.window_offscreen_cache.get(&node_id)?;
-    let texture = cache.texture.clone()?;
-    let ob = cache.bbox?;
-    if !cache.has_content {
-        return None;
-    }
-
-    let output_size = layer_output_size_for_monitor(st, monitor);
-    if output_size.w <= 0 || output_size.h <= 0 {
-        return None;
-    }
-    let output_clip = Rectangle::<i32, Physical>::new(
-        (0, 0).into(),
-        (output_size.w.max(1), output_size.h.max(1)).into(),
-    );
-
-    let (view_center, viewport_size, view_size) = render_view_for_monitor(st, monitor);
-    let render_scale = (viewport_size.x.max(1.0) / view_size.x.max(1.0)).max(0.01);
-    let local_geo = window_geometry_for_node(st, node_id).unwrap_or((
-        ob.loc.x as f32,
-        ob.loc.y as f32,
-        ob.size.w.max(1) as f32,
-        ob.size.h.max(1) as f32,
-    ));
-    let (cx, cy) = world_to_screen_for_view(
-        view_center,
-        view_size,
-        output_size.w,
-        output_size.h,
-        node.pos.x,
-        node.pos.y,
-    );
-    let gw = (local_geo.2 * render_scale).round().max(1.0) as i32;
-    let gh = (local_geo.3 * render_scale).round().max(1.0) as i32;
-    let gx = cx - (gw / 2);
-    let gy = cy - (gh / 2);
-    let fullscreen_on_monitor = st
-        .fullscreen_monitor_for_node(node_id)
-        .is_some_and(|fullscreen_monitor| fullscreen_monitor == monitor);
-    let effective_border_px = if fullscreen_on_monitor {
-        0
-    } else {
-        let base = st.runtime.tuning.border_size_px.max(0) as f32;
-        let scaled = (base * render_scale).round();
-        if base > 0.0 {
-            scaled.max(1.0) as i32
-        } else {
-            0
-        }
-    };
-    let effective_corner_radius_px = if fullscreen_on_monitor {
-        0
-    } else {
-        let base = st.runtime.tuning.border_radius_px.max(0) as f32;
-        let scaled = (base * render_scale).round();
-        if base > 0.0 {
-            scaled.max(1.0) as i32
-        } else {
-            0
-        }
-    };
-    let strict_square_csd_transition = strict_square_csd_transition_mode(
-        st.runtime.tuning.no_csd,
-        effective_corner_radius_px,
-        false,
-    );
-    let preserve_visual_margin = !strict_square_csd_transition
-        && !st.runtime.tuning.no_csd
-        && effective_corner_radius_px == 0;
-    let lock_dst_to_geometry = effective_corner_radius_px > 0;
-    let (src_x, src_y, src_w, src_h, dst_x, dst_y, dst_w, dst_h, clip_x, clip_y, clip_w, clip_h) =
-        offscreen_visual_crop_and_dst(
-            ob.loc.x,
-            ob.loc.y,
-            ob.size.w.max(1),
-            ob.size.h.max(1),
-            local_geo.0,
-            local_geo.1,
-            local_geo.2,
-            local_geo.3,
-            gx,
-            gy,
-            gw.max(1),
-            gh.max(1),
-            render_scale,
-            output_clip,
-            preserve_visual_margin,
-            lock_dst_to_geometry,
-        );
-    let src_scale_x = if src_w > 0.0 {
-        dst_w as f32 / src_w as f32
-    } else {
-        1.0
-    };
-    let src_scale_y = if src_h > 0.0 {
-        dst_h as f32 / src_h as f32
-    } else {
-        1.0
-    };
-    let disable_geo_clip = !strict_square_csd_transition
-        && !st.runtime.tuning.no_csd
-        && effective_corner_radius_px == 0;
-    let geo_local_x = local_geo.0 - ob.loc.x as f32;
-    let geo_local_y = local_geo.1 - ob.loc.y as f32;
-    let geo_src_x = (geo_local_x - src_x as f32).max(0.0);
-    let geo_src_y = (geo_local_y - src_y as f32).max(0.0);
-    let geo_offset_x = if disable_geo_clip {
-        0.0
-    } else {
-        (geo_src_x * src_scale_x).max(0.0)
-    };
-    let geo_offset_y = if disable_geo_clip {
-        0.0
-    } else {
-        (geo_src_y * src_scale_y).max(0.0)
-    };
-    let geo_w_px = if disable_geo_clip {
-        0.0
-    } else {
-        (local_geo.2 * src_scale_x).min(dst_w as f32).max(1.0)
-    };
-    let geo_h_px = if disable_geo_clip {
-        0.0
-    } else {
-        (local_geo.3 * src_scale_y).min(dst_h as f32).max(1.0)
-    };
-    let offscreen = OffscreenNodeTexture {
-        texture,
-        alpha: 1.0,
-        corner_radius: (effective_corner_radius_px - effective_border_px).max(0) as f32,
-        src_x,
-        src_y,
-        src_w,
-        src_h,
-        dst_x,
-        dst_y,
-        dst_w,
-        dst_h,
-        clip_x,
-        clip_y,
-        clip_w,
-        clip_h,
-        geo_offset_x,
-        geo_offset_y,
-        geo_w: geo_w_px,
-        geo_h: geo_h_px,
-    };
-    let border_rect = if effective_border_px > 0 {
-        let border_color = if st.model.focus_state.primary_interaction_focus == Some(node_id) {
-            let color = st.runtime.tuning.border_color_focused;
-            Color32F::new(color.r, color.g, color.b, 1.0)
-        } else {
-            let color = st.runtime.tuning.border_color_unfocused;
-            Color32F::new(color.r, color.g, color.b, 1.0)
-        };
-        Some(ActiveBorderRect {
-            x: gx,
-            y: gy,
-            w: gw.max(1),
-            h: gh.max(1),
-            inner_offset_x: effective_border_px as f32,
-            inner_offset_y: effective_border_px as f32,
-            inner_w: gw.max(1) as f32,
-            inner_h: gh.max(1) as f32,
-            alpha: 1.0,
-            border_px: effective_border_px as f32,
-            corner_radius: effective_corner_radius_px as f32,
-            inner_corner_radius: (effective_corner_radius_px - effective_border_px).max(0) as f32,
-            border_color,
-        })
-    } else {
-        None
-    };
-
-    Some((border_rect, vec![offscreen]))
-}
-
-pub(crate) fn prewarm_visible_active_window_offscreen_caches(
-    renderer: &mut GlesRenderer,
-    st: &mut Halley,
-    now: Instant,
-) {
-    let mut wl_surfaces: Vec<_> = st
-        .platform
-        .xdg_shell_state
-        .toplevel_surfaces()
-        .iter()
-        .filter_map(|toplevel| {
-            let wl = toplevel.wl_surface().clone();
-            let node_id = st.model.surface_to_node.get(&wl.id()).copied()?;
-            Some((node_id, wl))
-        })
-        .collect();
-
-    wl_surfaces.sort_by_key(|(id, _)| std::cmp::Reverse(id.as_u64()));
-
-    for (node_id, wl) in wl_surfaces {
-        let bbox = sync_node_size_from_surface(st, node_id, &wl);
-        let Some(node) = st.model.field.node(node_id) else {
-            continue;
-        };
-        if node.state != halley_core::field::NodeState::Active
-            || !st.model.field.is_visible(node_id)
-            || !st.node_visible_on_current_monitor(node_id)
-        {
-            continue;
-        }
-
-        let cache_missing = st
-            .ui
-            .render_state
-            .window_offscreen_cache
-            .get(&node_id)
-            .is_none_or(|cache| {
-                !cache.matches_size(bbox.size.w, bbox.size.h)
-                    || cache.texture.is_none()
-                    || cache.bbox.is_none()
-                    || !cache.has_content
-            });
-        if !cache_missing {
-            continue;
-        }
-
-        let cache = st.ui.render_state.ensure_window_offscreen_cache(
-            node_id,
-            bbox.size.w,
-            bbox.size.h,
-            now,
-        );
-        if !cache.dirty && cache.texture.is_some() && cache.bbox.is_some() && cache.has_content {
-            continue;
-        }
-
-        if let Ok(offscreen) = render_surface_tree_to_texture(renderer, &wl, 1.0, None) {
-            let cache = st
-                .ui
-                .render_state
-                .window_offscreen_cache
-                .get_mut(&node_id)
-                .expect("offscreen cache should exist after prewarm ensure");
-            cache.texture = Some(offscreen.texture);
-            cache.bbox = Some(offscreen.bbox);
-            cache.has_content = offscreen.has_content;
-            cache.mark_clean(now);
-        }
-    }
 }
 
 #[allow(clippy::type_complexity)]
@@ -1085,7 +484,8 @@ pub(crate) fn collect_active_surfaces(
                 .get(&node_id)
                 .copied()
         });
-        let transition_alpha = st.active_transition_alpha(node_id, now);
+        let transition_alpha =
+            crate::compositor::workspace::state::active_transition_alpha(st, node_id, now);
         let anim = crate::render::anim_style_for(st, node_id, node_state.clone(), now);
         let fullscreen_entry_scale = st.fullscreen_entry_scale(node_id, st.now_ms(now));
         let active_resize = active_resize_geometry_screen(st, node_id, resize_preview);
