+++ crates/halley-wl/src/render/window/offscreen.rs
@@ -1,0 +1,301 @@
+use super::*;
+
+pub(super) fn render_view_for_monitor(st: &Halley, monitor: &str) -> (Vec2, Vec2, Vec2) {
+    if st.model.monitor_state.current_monitor == monitor {
+        return (
+            st.model.viewport.center,
+            st.model.viewport.size,
+            st.model.zoom_ref_size,
+        );
+    }
+
+    st.model
+        .monitor_state
+        .monitors
+        .get(monitor)
+        .map(|space| {
+            (
+                space.viewport.center,
+                space.viewport.size,
+                space.zoom_ref_size,
+            )
+        })
+        .unwrap_or((
+            st.model.viewport.center,
+            st.model.viewport.size,
+            st.model.zoom_ref_size,
+        ))
+}
+
+pub(super) fn world_to_screen_for_view(
+    view_center: Vec2,
+    view_size: Vec2,
+    output_w: i32,
+    output_h: i32,
+    x: f32,
+    y: f32,
+) -> (i32, i32) {
+    let vw = view_size.x.max(1.0);
+    let vh = view_size.y.max(1.0);
+    let nx = ((x - view_center.x) / vw) + 0.5;
+    let ny = ((y - view_center.y) / vh) + 0.5;
+
+    (
+        (nx * output_w as f32).round() as i32,
+        (ny * output_h as f32).round() as i32,
+    )
+}
+
+pub(crate) fn capture_closing_window_animation(
+    st: &Halley,
+    monitor: &str,
+    node_id: NodeId,
+) -> Option<(Option<ActiveBorderRect>, Vec<OffscreenNodeTexture>)> {
+    let node = st.model.field.node(node_id)?;
+    let cache = st.ui.render_state.window_offscreen_cache.get(&node_id)?;
+    let texture = cache.texture.clone()?;
+    let ob = cache.bbox?;
+    if !cache.has_content {
+        return None;
+    }
+
+    let output_size = layer_output_size_for_monitor(st, monitor);
+    if output_size.w <= 0 || output_size.h <= 0 {
+        return None;
+    }
+    let output_clip = Rectangle::<i32, Physical>::new(
+        (0, 0).into(),
+        (output_size.w.max(1), output_size.h.max(1)).into(),
+    );
+
+    let (view_center, viewport_size, view_size) = render_view_for_monitor(st, monitor);
+    let render_scale = (viewport_size.x.max(1.0) / view_size.x.max(1.0)).max(0.01);
+    let local_geo = window_geometry_for_node(st, node_id).unwrap_or((
+        ob.loc.x as f32,
+        ob.loc.y as f32,
+        ob.size.w.max(1) as f32,
+        ob.size.h.max(1) as f32,
+    ));
+    let (cx, cy) = world_to_screen_for_view(
+        view_center,
+        view_size,
+        output_size.w,
+        output_size.h,
+        node.pos.x,
+        node.pos.y,
+    );
+    let gw = (local_geo.2 * render_scale).round().max(1.0) as i32;
+    let gh = (local_geo.3 * render_scale).round().max(1.0) as i32;
+    let gx = cx - (gw / 2);
+    let gy = cy - (gh / 2);
+    let fullscreen_on_monitor = st
+        .fullscreen_monitor_for_node(node_id)
+        .is_some_and(|fullscreen_monitor| fullscreen_monitor == monitor);
+    let effective_border_px = if fullscreen_on_monitor {
+        0
+    } else {
+        let base = st.runtime.tuning.border_size_px.max(0) as f32;
+        let scaled = (base * render_scale).round();
+        if base > 0.0 {
+            scaled.max(1.0) as i32
+        } else {
+            0
+        }
+    };
+    let effective_corner_radius_px = if fullscreen_on_monitor {
+        0
+    } else {
+        let base = st.runtime.tuning.border_radius_px.max(0) as f32;
+        let scaled = (base * render_scale).round();
+        if base > 0.0 {
+            scaled.max(1.0) as i32
+        } else {
+            0
+        }
+    };
+    let strict_square_csd_transition = strict_square_csd_transition_mode(
+        st.runtime.tuning.no_csd,
+        effective_corner_radius_px,
+        false,
+    );
+    let preserve_visual_margin = !strict_square_csd_transition
+        && !st.runtime.tuning.no_csd
+        && effective_corner_radius_px == 0;
+    let lock_dst_to_geometry = effective_corner_radius_px > 0;
+    let (src_x, src_y, src_w, src_h, dst_x, dst_y, dst_w, dst_h, clip_x, clip_y, clip_w, clip_h) =
+        offscreen_visual_crop_and_dst(
+            ob.loc.x,
+            ob.loc.y,
+            ob.size.w.max(1),
+            ob.size.h.max(1),
+            local_geo.0,
+            local_geo.1,
+            local_geo.2,
+            local_geo.3,
+            gx,
+            gy,
+            gw.max(1),
+            gh.max(1),
+            render_scale,
+            output_clip,
+            preserve_visual_margin,
+            lock_dst_to_geometry,
+        );
+    let src_scale_x = if src_w > 0.0 {
+        dst_w as f32 / src_w as f32
+    } else {
+        1.0
+    };
+    let src_scale_y = if src_h > 0.0 {
+        dst_h as f32 / src_h as f32
+    } else {
+        1.0
+    };
+    let disable_geo_clip = !strict_square_csd_transition
+        && !st.runtime.tuning.no_csd
+        && effective_corner_radius_px == 0;
+    let geo_local_x = local_geo.0 - ob.loc.x as f32;
+    let geo_local_y = local_geo.1 - ob.loc.y as f32;
+    let geo_src_x = (geo_local_x - src_x as f32).max(0.0);
+    let geo_src_y = (geo_local_y - src_y as f32).max(0.0);
+    let geo_offset_x = if disable_geo_clip {
+        0.0
+    } else {
+        (geo_src_x * src_scale_x).max(0.0)
+    };
+    let geo_offset_y = if disable_geo_clip {
+        0.0
+    } else {
+        (geo_src_y * src_scale_y).max(0.0)
+    };
+    let geo_w_px = if disable_geo_clip {
+        0.0
+    } else {
+        (local_geo.2 * src_scale_x).min(dst_w as f32).max(1.0)
+    };
+    let geo_h_px = if disable_geo_clip {
+        0.0
+    } else {
+        (local_geo.3 * src_scale_y).min(dst_h as f32).max(1.0)
+    };
+    let offscreen = OffscreenNodeTexture {
+        texture,
+        alpha: 1.0,
+        corner_radius: (effective_corner_radius_px - effective_border_px).max(0) as f32,
+        src_x,
+        src_y,
+        src_w,
+        src_h,
+        dst_x,
+        dst_y,
+        dst_w,
+        dst_h,
+        clip_x,
+        clip_y,
+        clip_w,
+        clip_h,
+        geo_offset_x,
+        geo_offset_y,
+        geo_w: geo_w_px,
+        geo_h: geo_h_px,
+    };
+    let border_rect = if effective_border_px > 0 {
+        let border_color = if st.model.focus_state.primary_interaction_focus == Some(node_id) {
+            let color = st.runtime.tuning.border_color_focused;
+            Color32F::new(color.r, color.g, color.b, 1.0)
+        } else {
+            let color = st.runtime.tuning.border_color_unfocused;
+            Color32F::new(color.r, color.g, color.b, 1.0)
+        };
+        Some(ActiveBorderRect {
+            x: gx,
+            y: gy,
+            w: gw.max(1),
+            h: gh.max(1),
+            inner_offset_x: effective_border_px as f32,
+            inner_offset_y: effective_border_px as f32,
+            inner_w: gw.max(1) as f32,
+            inner_h: gh.max(1) as f32,
+            alpha: 1.0,
+            border_px: effective_border_px as f32,
+            corner_radius: effective_corner_radius_px as f32,
+            inner_corner_radius: (effective_corner_radius_px - effective_border_px).max(0) as f32,
+            border_color,
+        })
+    } else {
+        None
+    };
+
+    Some((border_rect, vec![offscreen]))
+}
+
+pub(crate) fn prewarm_visible_active_window_offscreen_caches(
+    renderer: &mut GlesRenderer,
+    st: &mut Halley,
+    now: Instant,
+) {
+    let mut wl_surfaces: Vec<_> = st
+        .platform
+        .xdg_shell_state
+        .toplevel_surfaces()
+        .iter()
+        .filter_map(|toplevel| {
+            let wl = toplevel.wl_surface().clone();
+            let node_id = st.model.surface_to_node.get(&wl.id()).copied()?;
+            Some((node_id, wl))
+        })
+        .collect();
+
+    wl_surfaces.sort_by_key(|(id, _)| std::cmp::Reverse(id.as_u64()));
+
+    for (node_id, wl) in wl_surfaces {
+        let bbox = sync_node_size_from_surface(st, node_id, &wl);
+        let Some(node) = st.model.field.node(node_id) else {
+            continue;
+        };
+        if node.state != halley_core::field::NodeState::Active
+            || !st.model.field.is_visible(node_id)
+            || !st.node_visible_on_current_monitor(node_id)
+        {
+            continue;
+        }
+
+        let cache_missing = st
+            .ui
+            .render_state
+            .window_offscreen_cache
+            .get(&node_id)
+            .is_none_or(|cache| {
+                !cache.matches_size(bbox.size.w, bbox.size.h)
+                    || cache.texture.is_none()
+                    || cache.bbox.is_none()
+                    || !cache.has_content
+            });
+        if !cache_missing {
+            continue;
+        }
+
+        let cache = st.ui.render_state.ensure_window_offscreen_cache(
+            node_id,
+            bbox.size.w,
+            bbox.size.h,
+            now,
+        );
+        if !cache.dirty && cache.texture.is_some() && cache.bbox.is_some() && cache.has_content {
+            continue;
+        }
+
+        if let Ok(offscreen) = render_surface_tree_to_texture(renderer, &wl, 1.0, None) {
+            let cache = st
+                .ui
+                .render_state
+                .window_offscreen_cache
+                .get_mut(&node_id)
+                .expect("offscreen cache should exist after prewarm ensure");
+            cache.texture = Some(offscreen.texture);
+            cache.bbox = Some(offscreen.bbox);
+            cache.has_content = offscreen.has_content;
+            cache.mark_clean(now);
+        }
+    }
+}
